demo/services/cv/requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
pillow==10.0.1
easyocr==1.7.0
ultralytics==8.0.206
torch==2.1.0
numpy==1.24.3
